+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Changed
 
 - `demethodize` function now accepts object and name of the function to be extracted and returned function just accepts arguments of the original method.
+- `sequentially` returns function that accepts array of according function parameters now.
 
 ## [0.28.0] - 2021-07-22